math bug

+++ activationFunctions.py
@@ -16,7 +16,7 @@ class ActivationFunction:
 
     @staticmethod
     def swish(x: np.ndarray, dtype=np.float32):
-        return x.astype(dtype) / (1 + np.exp(-x, dtype=np.float64)).astype(dtype)
+        return (x / (1 + np.exp(-x, dtype=np.float64))).astype(dtype)
 
     @staticmethod
     def softmax(x: np.ndarray, dtype=np.float32):
@@ -39,7 +39,7 @@ class ActivationFunction:
 
     @staticmethod
     def d_swish(x: np.ndarray, dtype=np.float32):
-        sig = 1 / (1 + np.exp(-x, dtype=np.float64)).astype(dtype)
+        sig = (1 / (1 + np.exp(-x, dtype=np.float64))).astype(dtype)
         return sig * (1 + x.astype(dtype) * (1 - sig))
 
     @staticmethod
